% Generated by roxygen2 (4.0.1): do not edit by hand
\name{path_backward}
\alias{path_backward}
\title{Compute optimal u of Lemma 1 in JRSSB paper
using the backward algorithm}
\usage{
path_backward(L)
}
\arguments{
\item{L}{vector of eigenvalues}
}
\description{
Compute optimal u of Lemma 1 in JRSSB paper
using the backward algorithm
}

