% Generated by roxygen2 (4.0.1): do not edit by hand
\name{path_forward}
\alias{path_forward}
\title{Compute optimal u of Lemma 1 in JRSSB paper
using the forward algorithm}
\usage{
path_forward(L)
}
\arguments{
\item{L}{vector of eigenvalues}
}
\description{
Compute optimal u of Lemma 1 in JRSSB paper
using the forward algorithm
}

