% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condreg.R
\name{pfweights}
\alias{pfweights}
\title{Compute optimal portfolio weights}
\usage{
pfweights(sigma)
}
\arguments{
\item{sigma}{covariance matrix}
}
\value{
new portfolio weights
}
\description{
Compute optimal portfolio weights
}
