% Generated by roxygen2 (4.0.1): do not edit by hand
\name{transcost}
\alias{transcost}
\title{Compute transaction cost}
\usage{
transcost(wnew, wold, lastearnings, reltc, wealth)
}
\arguments{
\item{wnew}{new portfolio weights}

\item{wold}{old portfolio weights}

\item{lastearnings}{earnings from last period}

\item{reltc}{relative transaction cost}

\item{wealth}{current wealth}
}
\value{
transaction cost of rebalancing portfolio
}
\description{
Compute transaction cost
}

